src/requirements.txt
pandas
matplotlib
tqdm
opencv_python
albumentations
numpy
PyYAML
scikit_learn